requirements.txt
discord.py
llama-index
chromadb
python-dotenv
requests
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
sentence-transformers 